+++ CHANGELOG.md
@@ -16,6 +16,10 @@
   - feat(Input): implement `ref` API on input cell controls
   - feat(Select): implement `ref` API on select cell controls
   - feat(Autocompleter): implement `ref` API on autocompleter cell controls
+  - feat(CellControl): implement `className` API
+  - feat(Input): implement `className` API
+  - feat(Select): implement `className` API
+  - feat(Autocompleter): implement `className` API
 </details>
 
 ## 0.80.1